implement dumb backend
This backend is a stripped down version of "none" backend
without key pair generation / loading on the target machine.

It instead relies on system/user SSH config for both authentication
and host-key checking. This means that appropriate SSH keys needs to be added
to targets manually.

Related to #696 and many issues with SSH bootstrapping/recovery.